+++ src/Viserio/Bridge/Monolog/Tests/DataCollector/MonologLoggerDataCollectorTest.php
@@ -1,0 +1,3 @@
+<?php
+declare(strict_types=1);
+namespace Viserio\Bridge\Monolog\Tests\DataCollector;
