--- src/Viserio/Component/Profiler/Tests/DataCollector/Bridge/Monolog/MonologLoggerDataCollectorTest.php
@@ -1,3 +1,0 @@
-<?php
-declare(strict_types=1);
-namespace Viserio\Component\Profiler\Tests\DataCollector\Bridge;
